해시태그 두개가 안들어감

+++ src/main/java/flower/popupday/popup/dto/PopupHashTagDTO.java
@@ -4,22 +4,22 @@ import org.springframework.stereotype.Component;
 
 @Component("popupHashTagDTO")
 public class PopupHashTagDTO {
-    private int popup_id;
-    private int hash_tag_id;
+    private Long popup_id;
+    private Long hash_tag_id;
 
-    public int getPopup_id() {
+    public Long getPopup_id() {
         return popup_id;
     }
 
-    public void setPopup_id(int popup_id) {
+    public void setPopup_id(Long popup_id) {
         this.popup_id = popup_id;
     }
 
-    public int getHashtag_id() {
+    public Long getHashtag_id() {
         return hash_tag_id;
     }
 
-    public void setHashtag_id(int hash_tag_id) {
+    public void setHashtag_id(Long hash_tag_id) {
         this.hash_tag_id = hash_tag_id;
     }
 }
